+++ src/components/PortalsAndModals/PortalView.tsx
@@ -1,13 +1,14 @@
 import type { ScrollViewProps, StyleProp, ViewStyle } from 'react-native';
 import { Keyboard, Pressable, StyleSheet } from 'react-native';
 import { PageScrollView } from 'pagescrollview';
+import type { OmitKey } from '../../internalUtils/types';
 import { useTheme } from '../../main';
+import type { PortalProps } from './PortalsAndModals';
 import { Portal } from './PortalsAndModals';
 
 
 
-export type PortalViewProps = {
-  onCancel: () => void;
+export type PortalViewProps = OmitKey<PortalProps, 'style'> & {
   /** The style of the wrapping Portal' View. Useful to change the backdrop color or the ModalView flex position. */
   portalStyle?: StyleProp<ViewStyle>;
   /** Style of the wrapping box. */
@@ -24,14 +25,15 @@ export type PortalViewProps = {
 // TODO can use header and footer from PageScrollView, now a FlatList.
 // Actually it uses Footer. Maybe convert it to use children to allow us to use those two?
 export const PortalView: React.FC<PortalViewProps> = ({
-  onCancel, children,
+  children,
   containerStyle, portalStyle, contentStyle,
   scrollViewProps,
   Header, Footer,
+  ...portalProps
 }) => {
   const { colors } = useTheme();
   return (
-    <Portal onRequestClose={onCancel} viewStyle={portalStyle}>
+    <Portal style={portalStyle} {...portalProps}>
       <Pressable style={[s.container, { backgroundColor: colors.background }, containerStyle]} onPress={() => Keyboard.dismiss()}>
         {Header && (typeof Header === 'function' ? <Header/> : Header)}
         <PageScrollView viewStyle={[s.content, contentStyle]} {...scrollViewProps}>